requirements.txt
customtkinter
PyPDF2
python-dotenv
google-generativeai
google-api-core
requests
Pillow
